+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This changelog covers all three packages, as they are (for now) updated as a who
 ## Unreleased
 
 - Add File management views. Preview images and videos, download them. #121
+- Add `uploadFiles` method to @tomic/lib. #121
+- Add upload field to forms #121
 
 ## v0.29.2